Richness of a vector of individuals

Richness is the number of species.

# S3 method for character
Richness(
  NorP,
  Correction = "Chao1",
  Alpha = 0.05,
  JackOver = FALSE,
  ...,
  CheckArguments = TRUE
)

Arguments

NorP
A wmppp.object, with PointType values as species names.

Correction
A string containing one of the possible corrections: "None" (no correction), "Jackknife", "iChao1", or "Chao1", the default value.

Alpha
The risk level, 5% by default, used to optimize the jackknife order.

JackOver
If TRUE, retain the jackknife order immediately superior to the optimal one, usually resulting in the overestimation of the number of species. Default is FALSE.

...
Further arguments. Unsused.

CheckArguments
If TRUE (default), the function arguments are verified. Should be set to FALSE to save time in simulations for example, when the arguments have been checked elsewhere.

Value

A named number equal to the calculated diversity. The name is that of the bias correction used.

Examples

SpeciesList <- as.character(Paracou6$marks$PointType)
Richness(SpeciesList, Correction="Jackknife")

#> Jackknife 2 
#>         471