Species Distribution

A Species Distribution is a (preferably named) vector containing species abundances or probabilities. SpeciesDistribution objects include AbdVector and ProbaVector objects.

# S3 method for wmppp
as.SpeciesDistribution(x, ...)

# S3 method for factor
as.SpeciesDistribution(x, ...)

# S3 method for character
as.SpeciesDistribution(x, ...)

# S3 method for wmppp
as.ProbaVector(x, ...)

# S3 method for factor
as.ProbaVector(x, ...)

# S3 method for character
as.ProbaVector(x, ...)

# S3 method for wmppp
as.AbdVector(x, ...)

# S3 method for factor
as.AbdVector(x, ...)

# S3 method for character
as.AbdVector(x, ...)

Arguments

x

A wmppp.object with PointType values as species names, or a vector of factors or characters containing species names of each individual.

...

Further arguments. Unsused.

Value

A vector of species abundances (AbdVector) or probabilities (ProbaVector).

Details

as.AbdVector counts the number of individuals (points) per species (in marks$PointType).

as.ProbaVector normalizes the vector so that it sums to 1. If Correction is not "None", the observed abundance distribution is used to estimate the actual species distribution. The list of species will be changed: zero-abundance species will be cleared, and some unobserved species will be added. First, observed species probabilities are estimated folllowing Chao and Shen (2003) , i.e. input probabilities are multiplied by the sample coverage, or according to more sophisticated models: Chao et al. (2013) , single-parameter model or Chao and Jost (2015) , two-parameter model. The total probability of observed species equals the sample coverage. Then, the distribution of unobserved species can be unveiled: their number is estimated according to RCorrection (if the Jackknife estimator is chosen, the JackOver argument allows using the order immediately over the optimal one). The coverage deficit (1 minus the sample coverage) is shared by the unobserved species equally: Unveiling = "unif", Chao et al. (2013) or according to a geometric distribution: Unveiling = "geom", Chao and Jost (2015) .

SpeciesDistribution objects can be plotted. The plot method returns the estimated parameters of the fitted distribution. The broken stick has no parameter, so the maximum abundance is returned.

References

Chao A, Jost L (2015). “Estimating Diversity and Entropy Profiles via Discovery Rates of New Species.” Methods in Ecology and Evolution, 6(8), 873--882. doi: 10.1111/2041-210X.12349 .

Chao A, Shen T (2003). “Nonparametric Estimation of Shannon's Index of Diversity When There Are Unseen Species in Sample.” Environmental and Ecological Statistics, 10(4), 429--443. doi: 10.1023/A:1026096204727 .

Chao A, Wang Y, Jost L (2013). “Entropy and the Species Accumulation Curve: A Novel Entropy Estimator via Discovery Rates of New Species.” Methods in Ecology and Evolution, 4(11), 1091--1100. doi: 10.1111/2041-210x.12108 .