Tsallis (HCDT) Entropy of a vector of individuals

Calculates the HCDT, also known as Tsallis entropy of order $q$ of a vector of individuals

# S3 method for factor
Tsallis(NorP, q = 1, Correction = "Best", ..., CheckArguments = TRUE)

Arguments

NorP

A vector of factors.

q

A number: the order of entropy. Some corrections allow only a positive number. Default is 1 for Shannon entropy.

Correction

A string containing one of the possible corrections: "None" (no correction), "ChaoShen", "GenCov", "Grassberger", "Holste", "Bonachela", "ZhangGrabchak", or "ChaoWangJost", "Marcon", "UnveilC", "UnveiliC", "UnveilJ" or "Best", the default value. Currently, "Best" is "ChaoWangJost".

...

Further arguments. Unsused.

CheckArguments

If TRUE (default), the function arguments are verified. Should be set to FALSE to save time in simulations for example, when the arguments have been checked elsewhere.

Value

A named number equal to the calculated entropy. The name is that of the bias correction used.

Details

Tsallis (Havrda and Charvát 1967; Daróczy 1970; Tsallis 1988) generalized entropy is a generalized measure of diversity (Jost 2006) . See Tsallis for more details.

References

Daróczy Z (1970). “Generalized Information Functions.” Information and Control, 16(1), 36--51. doi: 10.1016/s0019-9958(70)80040-7 , https://doi.org/10.1016/s0019-9958(70)80040-7.

Havrda J, Charvát F (1967). “Quantification Method of Classification Processes. Concept of Structural Alpha-Entropy.” Kybernetika, 3(1), 30--35.

Jost L (2006). “Entropy and Diversity.” Oikos, 113(2), 363--375. doi: 10.1111/j.2006.0030-1299.14714.x .

Tsallis C (1988). “Possible Generalization of Boltzmann-Gibbs Statistics.” Journal of Statistical Physics, 52(1), 479--487. doi: 10.1007/BF01016429 .

Examples

Tsallis(Paracou6$marks$PointType)
#>  UnveilJ 
#> 4.702037